Use skia's support for parsing and rendering SVG
Don't use lafriks/go-svg to parse the SVG source. Don't construct
skia paths manually. Don't render each of the paths. Instead use
skia's SVG support to parse the svg and generate a DOM. Then use skia
to draw the DOM to the canvas.

This means that more svg features (such as dashes, and gradients) are
naturally supported.

The one major downside to this is that it removes the ability to
supply a Paint for drawing, for example to override paths' color.